+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/GameObjectBoundsInstantiator.cs
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2023 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Scripting.APIUpdating;
@@ -42,8 +41,7 @@ namespace Unity.Cloud.Gltfast
         }
 
         /// <inheritdoc />
-        [Obsolete("Use IInstantiator.AddMesh instead.")]
-        public override void AddPrimitive(
+        public override void AddMesh(
             uint nodeIndex,
             string meshName,
             MeshResult meshResult,
@@ -53,7 +51,7 @@ namespace Unity.Cloud.Gltfast
             int meshNumeration = 0
         )
         {
-            base.AddPrimitive(
+            base.AddMesh(
                 nodeIndex,
                 meshName,
                 meshResult,